+++ modules/core/pom.xml
@@ -211,9 +211,9 @@
 
 	<dependencies>
 		<dependency>
-			<groupId>com.opencsv</groupId>
-			<artifactId>opencsv</artifactId>
-		</dependency>
+            <groupId>au.com.bytecode.opencsv.wso2</groupId>
+            <artifactId>opencsv</artifactId>
+        </dependency>
 		<dependency>
 			<groupId>org.wso2.securevault</groupId>
 			<artifactId>org.wso2.securevault</artifactId>
