+++ modules/core/src/main/java/org/apache/synapse/mediators/util/collectors/JsonArrayCollector.java
@@ -22,7 +22,8 @@ package org.apache.synapse.mediators.util.collectors;
 import com.google.common.collect.Sets;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
-import org.apache.synapse.mediators.util.SimpleMessageContext;
+import com.google.gson.JsonObject;
+import org.apache.synapse.mediators.SimpleMessageContext;
 
 import java.util.Set;
 import java.util.function.BiConsumer;
@@ -36,13 +37,22 @@ import static java.util.stream.Collector.Characteristics.UNORDERED;
 /**
  * Collector to collect a Stream of JsonElement to a Json array and set the payload content
  */
-public class JsonArrayCollector implements Collector<JsonElement, JsonArray, Boolean> {
+public class JsonArrayCollector implements Collector<JsonElement, JsonArray, JsonElement> {
 
     private final SimpleMessageContext simpleMessageContext;
+    private final String objectKey;
 
-    public JsonArrayCollector(SimpleMessageContext simpleMessageContext) {
+    /**
+     * Create new instance with MessageConetext and Object key
+     *
+     * @param simpleMessageContext SimpleMessageContext to use
+     * @param objectKey            If this is not null, a json object is created and inside that object this key would be added
+     *                             with the value generated by the collector. If null, then JsonArray would be set as payload
+     */
+    public JsonArrayCollector(SimpleMessageContext simpleMessageContext, String objectKey) {
 
         this.simpleMessageContext = simpleMessageContext;
+        this.objectKey = objectKey;
     }
 
     @Override
@@ -74,11 +84,20 @@ public class JsonArrayCollector implements Collector<JsonElement, JsonArray, Boo
     }
 
     @Override
-    public Function<JsonArray, Boolean> finisher() {
+    public Function<JsonArray, JsonElement> finisher() {
 
-        return result -> {
-                simpleMessageContext.setJsonPayload(result);
-            return true;
+        return jsonArray -> {
+            JsonElement result;
+            if (objectKey == null) {
+                result = jsonArray;
+            } else {
+                JsonObject rootObject = new JsonObject();
+                rootObject.add(objectKey, jsonArray);
+                result = rootObject;
+            }
+            simpleMessageContext.setJsonPayload(result);
+
+            return result;
         };
     }
